fix layout some
