--- __init__.py
@@ -1,3 +1,0 @@
-from .dspac_api import DSPACAPI
-
-__all__ = ["DSPACAPI"]
