+++ dspac_api/__init__.py
@@ -1,0 +1,3 @@
+from .dspac_api import DSPACAPI
+
+__all__ = ["DSPACAPI"]
